PythonAppiumClient/requirements.txt
Appium-Python-Client==0.51
pytest==5.4.1
py==1.8.1
